+++ package.json
@@ -12,7 +12,9 @@
     "@react-navigation/native": "^5.1.6",
     "@react-navigation/stack": "^5.2.13",
     "expo": "~37.0.3",
+    "expo-font": "~8.1.0",
     "expo-linear-gradient": "~8.1.0",
+    "lottie-react-native": "~2.6.1",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
